(equipment) Fix field size of brand user and image counts

+++ astrobin_apps_equipment/models/equipment_brand.py
@@ -62,12 +62,12 @@ class EquipmentBrand(SafeDeleteModel):
     # This items are synced back from the search index.                                                                #
     ####################################################################################################################
 
-    user_count = models.PositiveSmallIntegerField(
+    user_count = models.PositiveIntegerField(
         editable=False,
         default=0
     )
 
-    image_count = models.PositiveSmallIntegerField(
+    image_count = models.PositiveIntegerField(
         editable=False,
         default=0
     )
